+++ crates/astria-core/src/protocol/account/mod.rs
@@ -1,3 +1,1 @@
 pub mod v1;
-
-use crate::generated::protocol::accounts::v1 as raw;