+++ crates/astria-core/CHANGELOG.md
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added method `TracePrefixed::leading_channel` to read the left-most channel of
   a trace prefixed ICS20 asset [#1768](https://github.com/astriaorg/astria/pull/1768)
 
+### Changed
+
+- Move all Astria APIs generated from the Protobuf spec from `astria_core::generated`
+  to `astria_core::generated::astria`
+  [#1825](https://github.com/astriaorg/astria/pull/1825)
+
 ### Removed
 
 - Removed method `TracePrefixed::last_channel` [#1768](https://github.com/astriaorg/astria/pull/1768)